pyproject.toml
[tool.poetry]
name = "cake"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bitarray = "^2.9.2"
kyber-py = "^0.2.2"
pycryptodome = "^3.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
